This commit speeds up the most time-consuming part of applying corrections. For a single call to `apply_all_corrections`:

Time spent in get_normalization_factor: 0.0007276535034179688 s
Time spent in maps_coefficient_getter: 0.00017762184143066406 s
Time spent in maps_coefficient_getter: 0.00012445449829101562 s
Time spent in apply_all_correction_single_maps: 0.0012238025665283203 s
**Time spent in get_maps_coefficient: 27.98320436477661 s**
Time spent in time_coefs_corr: 0.024751663208007812 s
Time spent in correct_geometry_: 0.003904581069946289 s
**Time spent in get_maps_coefficient: 25.5330171585083 s**
Time spent in time_coefs_corr: 0.018788576126098633 s
Time spent in correct_lifetime_: 0.018901586532592773 s

The new implementation takes advantage of numpy's internal C-speed loops to retrieve the data faster.

Using the trigger1 merged kdst for run 6971 I get this numbers:
Before: Time spent in get_maps_coefficient: 27.98320436477661 s
After : Time spent in get_maps_coefficient: 0.324688196182251 s